+++ libraries/helpers.rb
@@ -22,6 +22,21 @@ module Chefwaiter
     ::Chef.node['os'] == name
   end
 
+  def self.chef_version
+    # chef on windows has a different name than linux.
+    if os? 'windows'
+      chef_package = ::Chef.node['packages'].select do |pkg_name|
+        pkg_name =~ /[Cc]hef/
+      end
+      # We get back an array with a name that we don't know as it
+      # has the version number in it.
+      # Therefore we need to pull whatever we got out and use that.
+      return chef_package[chef_package.keys.first]['version']
+    end
+
+    ::Chef.node['packages']['chef']['version']
+  end
+
   # These platforms require upstart to run chef-waiter
   def self.upstart_platforms?
     return true if ::Chef.node['platform'] == 'amazon'
@@ -73,9 +88,7 @@ module Chefwaiter
   end
 
   def self.binary_location
-    s = ::File.join(::Chef.node['chef-waiter']['binary_path'], ::Chef.node['chef-waiter']['exec_name'])
-    s += '.exe' if os?('windows')
-    s
+    ::File.join(::Chef.node['chef-waiter']['binary_path'], ::Chef.node['chef-waiter']['exec_name'])
   end
 
   def self.service_version
@@ -84,11 +97,7 @@ module Chefwaiter
 
     # Chef 12 and 13 users will still use to the old code.
     # In chef 14 it ShellOut was remove.
-    if ::Chef.node['packages']['chef']['version'] < "14"
-      cmd = ::Chef::Mixin::ShellOut.shell_out(cmdstr)
-    else
-      cmd = ::Chef::Mixin::ShellOut.shell_out_command(cmdstr)
-    end
+    cmd = chef_version < '14' ? ::Chef::Mixin::ShellOut.shell_out(cmdstr) : ::Chef::Mixin::ShellOut.shell_out_command(cmdstr)
     cmd.stdout.chomp
   end