+++ test/integration/windows/serverspec/service_spec.rb
@@ -9,3 +9,7 @@ describe file('c:\\Program Files\\chefwaiter\\config.json') do
   it { should be_file }
   its(:content) { should match /"state_table_size": 40,/ }
 end
+
+describe file('c:\\Program Files\\chefwaiter\\chef-waiter.exe') do
+  it { should be_file }
+end