+++ CHANGELOG.md
@@ -5,3 +5,10 @@ This file is used to list changes made in each version of the chef-waiter cookbo
 
 1.0.1
 -------
+
+Inital public verion
+
+1.1.0
+-------
+
+Adding in better support for windows and updating the metadata file